Remove typo test file
